% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{convert_lat_lon}
\alias{convert_lat_lon}
\title{Convert coordinates to numerics}
\usage{
convert_lat_lon(df)
}
\arguments{
\item{df}{dataframe containing raw NOAA data}
}
\description{
Convert latitude and longitudes to numerics and renames them for compatibility with leaflet.
}
\seealso{
\code{\link[=eq_clean_data]{eq_clean_data()}}
}
