% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Clean NOAA Earthquake Data}
\usage{
eq_clean_data(df)
}
\arguments{
\item{df}{a dataframe or tibble containing raw NOAA data}
}
\value{
a cleaned dataframe
}
\description{
Prepares the NOAA earthquake data for use by downstream functions in this library.
}
\details{
This function cleans the columns of the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA earthquake data}
for further use with this library's plotting functions. This function performs the following:
\itemize{
\item consolidates the month, day and year columns into a single datetime column
\item separates the country and location into two separate columns, trims whitespace and converts them to titlecase
\item converts latitude and longitude into numerics and renames these columns for compatibility with Leaflet
}
}
\examples{
\dontrun{
 data <- eq_clean_data(raw_data)
}

}
