% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{eq_create_label}
\alias{eq_create_label}
\title{Create an HTML label column}
\usage{
eq_create_label(df)
}
\arguments{
\item{df}{dataframe containing NOAA earthquake data}
}
\value{
a character vector containing the labels
}
\description{
Creates a new column from a dataframe containing NOAA earthquake data
}
\details{
The function expects that the following named columns are present:
\itemize{
\item Location Name
\item Mag
\item Deaths
}
The values of these columns are used to create a new column containing
the HTML labels that is returned
}
\examples{
\dontrun{
df \%>\%
 mutate(labels = eq_create_label(.))
}

}
\seealso{
Other Mapping: 
\code{\link{eq_map}()},
\code{\link{html_label}()}
}
\concept{Mapping}
