% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Create location columns}
\usage{
eq_location_clean(data)
}
\arguments{
\item{data}{dataframe containing raw NOAA earthquake data}
}
\value{
dataframe with an additional "Country Name" column.
}
\description{
Splits location name column into separate country and location columns.
}
\details{
This function splits the "Location Name" into a "Country Name" column containing
the country and replaces the contents of "Location Name" with just the location.
}
\seealso{
\code{\link[=eq_clean_data]{eq_clean_data()}}, \code{\link[=country_name]{country_name()}}, \code{\link[=clean_location]{clean_location()}}
}
