% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{eq_map}
\alias{eq_map}
\title{Create a map of earthquake locations}
\usage{
eq_map(df, annot_col = NULL)
}
\arguments{
\item{df}{dataframe containing NOAA earthquake data}

\item{annot_col}{optional, a column label to use for annotating the earthquake markers}
}
\value{
a leaflet map object
}
\description{
Create a leaflet map showing the locations of earthquakes
}
\examples{
\dontrun{
  m <- df \%>\%
   mutate(label = eq_create_labels(.)) \%>\%
   eq_map(annot_col="label")
  m
}
}
\seealso{
Other Mapping: 
\code{\link{eq_create_label}()},
\code{\link{html_label}()}
}
\concept{Mapping}
