% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_labels.R
\name{geom_label_timeline}
\alias{geom_label_timeline}
\title{Add labels to timeline}
\usage{
geom_label_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by \code{aes()}}

\item{data}{The data to be displayed in this layer}
}
\description{
Add labels to a timeline geom with option to limit to top n largest earthquakes
}
\details{
This geom adds labels to the figure generated by \code{\link[=geom_timeline]{geom_timeline()}} and must be used
on conjunction with this geom as it inherits several of its aesthetics.
}
\section{Aesthetics}{


The following aesthetics are supported. Position aesthetics are inherited from \code{\link[=geom_timeline]{geom_timeline()}}.

\describe{
\item{label}{character with the text of the label}
\item{n_max}{optional integer specifying the top number of earthquakes to display. The \code{size} aesthetic inherited from \code{\link[=geom_timeline]{geom_timeline()}} is used to sort the earthquakes.}
}
}

\examples{
\dontrun{
data <- eq_clean(raw_data)
 data \%>\%
  ggplot(aes(x=Year, y=Country, size=Magnitude, fill=Deaths)) +
  geom_timeline(alpha=0.5) +
  geom_label_timeline(aes(label=label), x_max=5)
}


}
\seealso{
\code{\link[=geom_timeline]{geom_timeline()}}

Other Timeline Geom Label: 
\code{\link{GeomLabelTimeline}}
}
\concept{Timeline Geom Label}
