% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_geom.R
\name{geom_timeline}
\alias{geom_timeline}
\title{Create an earthquake timeline}
\usage{
geom_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by \code{aes()}}

\item{data}{The data to be displayed in this layer}
}
\description{
Creates a timeline of earthquakes for each specified country passed to the geom.
}
\details{
This geom is for the visualization of the \href{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}{NOAA earthquake data}. The
geom displays a timeline where circles mark the year of an earthquake and whose radius is proportional
to the magnitude and whose color fill intensity is proportional to the number of deaths. A factor can be supplied
to the geom whose levels are countries and a separate timeline will be draw for each country. It is recommended
that no more than three countries be visualized in this way to avoid crowding the figure.

This geom is not to be used with the raw data from NOAA but must be first cleaned with the \code{\link[=eq_clean_data]{eq_clean_data()}} function.
}
\section{Aesthetics}{


The \code{geom_timeline()} understands the following aesthetics
\describe{
\item{x}{numeric representing the year of the earthquake}
\item{y}{factor representing the region. A new timeline will be plotted for each level of the factor}
\item{size}{a numeric controlling the size of the circle marks along the timeline}
\item{fill}{a numeric controlling the fill of the circle marks}
\item{alpha}{optional numeric between 0 and 1 controlling the transparency of the circles}
\item{color}{optional character specifying the color of the circle borders}
}
}

\examples{
\dontrun{
 data <- eq_clean(raw_data)
 data \%>\%
  ggplot(aes(x=Year, y=Country, size=Magnitude, fill=Deaths)) +
  geom_timeline(alpha=0.5)
}

}
\seealso{
\code{\link[=eq_clean_data]{eq_clean_data()}}, \code{\link[=geom_label_timeline]{geom_label_timeline()}}
}
\concept{Geom Timeline}
