% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_labels.R
\docType{data}
\name{GeomLabelTimeline}
\alias{GeomLabelTimeline}
\title{Base class for geom_label_timeline}
\format{
An object of class \code{GeomLabelTimeline} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomLabelTimeline
}
\description{
Provides the baseclass for the \code{\link[=geom_label_timeline]{geom_label_timeline()}} geom. This function is not
meant to be called directly by the user and is not exported. Please see \code{\link[=geom_label_timeline]{geom_label_timeline()}}.
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::Geom()}}

Other Timeline Geom Label: 
\code{\link{geom_label_timeline}()}
}
\concept{Timeline Geom Label}
\keyword{datasets}
