% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeline_geom.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{Base class for geom_timeline}
\format{
An object of class \code{GeomTimeline} (inherits from \code{GeomPoint}, \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTimeline
}
\description{
Provides the baseclass for the \code{\link[=geom_timeline]{geom_timeline()}} geom. This function is not
meant to be called directly by the user and is not exported. Please see \code{\link[=geom_timeline]{geom_timeline()}}
}
\seealso{
\code{\link[ggplot2:ggplot2-ggproto]{ggplot2::Geom()}}
}
\concept{Timeline Geom}
\keyword{datasets}
