% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_utils.R
\name{html_label}
\alias{html_label}
\title{Create HTML labels}
\usage{
html_label(location, mag, deaths)
}
\arguments{
\item{location}{string containing the name of earthquake location}

\item{mag}{numeric with the earthquake magnitude}

\item{deaths}{numeric with the number of deaths}
}
\value{
string containing the HTML label
}
\description{
Create HTML labels for earthquake leaflet map using fields from NOAA data
}
\details{
Given a location, magnitude and number of deaths, this function returns a formatted
HTML string for use with Leaflet. If any of the values are NA, that value will not
be included in the label.
}
\seealso{
Other Mapping: 
\code{\link{eq_create_label}()},
\code{\link{eq_map}()}
}
\concept{Mapping}
