% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_to_interval}
\alias{map_to_interval}
\title{Linear transformation of an interveral}
\usage{
map_to_interval(x, x_max, x_min, u = 1, l = 0)
}
\arguments{
\item{x}{a numeric, point to be scaled}

\item{x_max}{upper bound of the domain}

\item{x_min}{lower bound of the domain}

\item{u}{upper bound of the range}

\item{l}{lower bound of the range}
}
\value{
numeric scaled between upper and lower interval bounds
}
\description{
Map points in the interval [x_max, x_min] to the interval [l, u].
}
