% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processing.R
\name{parse_date}
\alias{parse_date}
\title{Format NOAA dates}
\usage{
parse_date(data)
}
\arguments{
\item{data}{dataframe containing raw NOAA earthquake data}
}
\value{
a dataframe with a "date" column added and "Year", "Mo" and "Dy" columns removed.
}
\description{
Convert year, month and day columns into a single datetime column.
}
\seealso{
\code{\link[=eq_clean_data]{eq_clean_data()}}
}
