% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_noaa_sample}
\alias{raw_noaa_sample}
\title{Sample of NOAA earthquake data}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 287 rows and 38 columns.
}
\source{
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
raw_noaa_sample
}
\description{
This is a sample of the raw NOAA earthquake containing earthquakes from Mexico and California
}
\examples{
data("raw_noaa_sample")

}
\keyword{datasets}
