% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{test_data_1}
\alias{test_data_1}
\title{Data used for unit tests}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 38 columns.
}
\source{
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/search}
}
\usage{
test_data_1
}
\description{
Contains the 100th row of \code{raw_noaa_sample} data and is used for unit testing.
}
\keyword{datasets}
