% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_timeline_labels.R
\name{geom_label_timeline}
\alias{geom_label_timeline}
\title{Add labels to timeline}
\usage{
geom_label_timeline(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{A set of aesthetic mappings created by \code{aes()}}

\item{data}{The data to be displayed in this layer}

\item{stat}{The statistical transformation to use on the data for this
layer, as a string.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{na.rm}{If \code{FALSE}, the default, missing values are removed with
a warning. If \code{TRUE}, missing values are silently removed.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{...}{Other arguments passed on to \code{\link[ggplot2:layer]{layer()}}. These are
often aesthetics, used to set an aesthetic to a fixed value, like
\code{colour = "red"} or \code{size = 3}. They may also be parameters
to the paired geom/stat.}
}
\description{
Add labels to a timeline geom with option to limit to top n largest earthquakes
}
\details{
This geom adds labels to the figure generated by \code{\link[=geom_timeline]{geom_timeline()}} and must be used
on conjunction with this geom as it inherits several of its aesthetics.
}
\section{Aesthetics}{


The following aesthetics are supported. Position aesthetics are inherited from \code{\link[=geom_timeline]{geom_timeline()}}.

\describe{
\item{label}{character with the text of the label}
\item{n_max}{optional integer specifying the top number of earthquakes to display. The \code{size} aesthetic inherited from \code{\link[=geom_timeline]{geom_timeline()}} is used to sort the earthquakes.}
}
}

\examples{
\dontrun{
data <- eq_clean(raw_data)
 data \%>\%
  ggplot(aes(x=Year, y=Country, size=Magnitude, fill=Deaths)) +
  geom_timeline(alpha=0.5) +
  geom_label_timeline(aes(label=label), x_max=5)
}

}
\seealso{
\code{\link[=geom_timeline]{geom_timeline()}}

Other Timeline Geom Label: 
\code{\link{GeomLabelTimeline}}
}
\concept{Timeline Geom Label}
